Opt: WebSocket integration for real-time updates
