+++ desktop.js
@@ -9,35 +9,7 @@
 // DOM 要素参照は後で初期化するためのプレースホルダ
 let iconShapeSelector = null;
 
-// ビルトインアイコン (index.html にある `appicon-...`) の既定 URL マップ
-const builtinIconUrls = {
-  'appicon-chrome': 'chrome://newtab',
-  'appicon-files': 'chrome://file-manager',
-  'appicon-settings': 'chrome://os-settings'
-};
-
-// localStorageで使用するキーを一元管理
-const LS_KEYS = {
-  WIDGET_POSITIONS: 'widgetPositions',
-  CUSTOM_APPS: 'customApps',
-  LINUX_APPS: 'linuxApps',
-  FILE_SHORTCUTS: 'fileShortcuts',
-  FOLDER_SHORTCUTS: 'folderShortcuts',
-  APP_FOLDERS: 'appFolders',
-  LANGUAGE: 'language',
-  GRID_MODE_ENABLED: 'gridModeEnabled',
-  ICON_SHAPE: 'iconShape',
-  COLOR_SCHEME: 'colorScheme',
-  CUSTOM_COLOR: 'customColor',
-  SHOW_SETTINGS_FAB: 'showSettingsFab',
-  BLUR_EFFECT_ENABLED: 'blurEffectEnabled',
-  DARK_MODE_ENABLED: 'darkModeEnabled',
-  DARK_MODE_SETTING: 'darkModeSetting',
-  GITHUB_USERNAME: 'githubUsername',
-  GOOGLE_CALENDAR_URL: 'googleCalendarUrl',
-  WIDGET_VISIBILITY: 'widgetVisibility',
-  WIDGET_SIZE_PREFIX: 'widgetSize:',
-};
+// Constants are defined in constants.js
 
 // ========================================
 // ダイアログヘルパー関数
@@ -106,7 +78,7 @@ async function showConfirmDialog(message, title = '', options = {}) {
   const okBtn = document.getElementById('global_dialog_ok');
   
   const lang = getCurrentLanguage();
-  cancelBtn.textContent = translations[lang].cancel || 'Cancel';
+  cancelBtn.textContent = i18n.t('cancel');
   okBtn.textContent = 'OK';
   
   dialogTitle.textContent = title;
@@ -412,7 +384,7 @@ async function updateMediaPlayer() {
   
   if (!hasPlayer) {
     const lang = getCurrentLanguage();
-    if (mediaTitle) mediaTitle.textContent = translations[lang].no_music_playing;
+    if (mediaTitle) mediaTitle.textContent = i18n.t('no_music_playing');
     if (mediaArtist) mediaArtist.textContent = '';
     if (mediaArt) mediaArt.style.display = 'none';
     if (mediaPlayIcon) mediaPlayIcon.setAttribute('name', 'play_arrow');
@@ -644,10 +616,12 @@ if (mediaSeekbar) {
 }
 
 // ブラウザからのメディア更新を受信
-window.onBrowserMediaUpdate = (info) => {
-  browserMediaInfo = info;
-  updateMediaPlayer();
-};
+if (window.electronAPI && window.electronAPI.onBrowserMediaUpdate) {
+  window.electronAPI.onBrowserMediaUpdate((info) => {
+    browserMediaInfo = info;
+    updateMediaPlayer();
+  });
+}
 
 // 定期的にメディア情報を更新（2秒ごと）
 setInterval(updateMediaPlayer, 1000);
@@ -1570,7 +1544,7 @@ async function handleFolderItemClick(app, modal) {
     const result = await openFileOrFolder(app.path);
     if (!result.success) {
       const lang = getCurrentLanguage();
-      await showAlertDialog(translations[lang].open_failed + ': ' + result.error);
+      await showAlertDialog(i18n.t('open_failed') + ': ' + result.error);
     }
     return;
   }
@@ -2571,243 +2545,7 @@ document.getElementById('refresh_page').onclick = () => {
   location.reload();
 }
 
-const translations = {
-  ja: {
-    files: "ファイル",
-    settings: "設定",
-    add_app: "アプリを追加",
-    add_widget: "ウィジェットを追加",
-    upload_image: "アプリ画像をアップロード",
-    app_name: "アプリ名",
-    url: "URL",
-    close: "とじる",
-    add: "追加",
-    change_position: "位置の変更",
-    refresh_dev: "リフレッシュ(開発者向け)",
-    grid_mode: "グリッドモード",
-    grid_mode_on: "ON",
-    grid_mode_off: "OFF",
-    save: "保存",
-    reset_position: "位置をリセット",
-    general: "一般",
-    keyboard_shortcuts: "キーボードショートカット",
-    design_style: "デザインとスタイル",
-    security_privacy: "セキュリティとプライバシー",
-    account: "アカウント",
-    debug: "デバッグ",
-    dev_settings: "開発者向け設定",
-    about: "バージョン情報",
-    user_agent: "ユーザーエージェント:",
-    language: "言語",
-    settings_button_label: "設定ボタン",
-    show_settings_button: "表示",
-    hide_settings_button: "非表示",
-    target_display: "ターゲットディスプレイ",
-    select_app_type: "追加するアプリの種類を選択",
-    web_app: "Webアプリ / URL",
-    linux_app: "Linuxアプリ",
-    add_linux_app: "Linuxアプリを追加",
-    command: "コマンド",
-    run_in_terminal: "ターミナルで実行",
-    blur_effect: "ブラー効果",
-    enabled: "有効",
-    disabled: "無効",
-    dark_mode: "ダークモード",
-    edit: "編集",
-    delete: "削除",
-    hide_widget: "ウィジェットを非表示",
-    edit_app: "アプリを編集",
-    edit_linux_app: "Linuxアプリを編集",
-    change_image: "画像を変更",
-    add_web_app: "アプリを追加",
-    confirm_delete: "を削除しますか？",
-    confirm_remove_from_folder: "をフォルダーから取り出しますか？",
-    confirm_reset_position: "すべてのウィジェットとアイコンの位置をリセットしますか？",
-    enter_name_and_url: "名前とURLを入力してください",
-    enter_name_and_command: "名前とコマンドを入力してください",
-    launch_failed: "アプリの起動に失敗しました",
-    enter_folder_name: "フォルダー名を入力",
-    color_scheme: "カラースキーム",
-    delete_all_data: "すべてのデータを削除",
-    confirm_delete_all_data: "すべてのデータ（アプリ、フォルダ、設定など）を削除しますか？\nこの操作は取り消せません。",
-    data_deleted: "すべてのデータを削除しました。ページを再読み込みします。",
-    file_shortcut: "ファイルのショートカット",
-    folder_shortcut: "フォルダのショートカット",
-    open_failed: "ファイル/フォルダを開けませんでした",
-    icon_shape: "アイコンの形",
-    square: "正方形",
-    circle: "円形",
-    window_count: "ウィンドウ数（仮想デスクトップ用）",
-    apply_restart: "適用して再起動",
-    confirm_restart: "ウィンドウ数を変更するにはアプリを再起動する必要があります。再起動しますか？",
-    folder_style: "フォルダーのスタイル",
-    apply_to_all: "すべてに適用",
-    google_login: "Googleアカウントにログイン",
-    google_login_help: "カレンダーが表示されない場合はログインしてください",
-    icon_settings: "アイコン設定",
-    system_settings: "システム設定",
-    data_management: "データ管理",
-    folder_settings: "フォルダー設定",
-    folder_name: "フォルダー名",
-    background_color: "背景色",
-    opacity: "透明度",
-    extract_color_from_image: "画像から色を抽出",
-    upload_image_generic: "画像をアップロード",
-    no_music_playing: "再生中の音楽なし",
-    loading: "読み込み中...",
-    set_username: "設定ボタンからユーザー名を設定してください",
-    github_settings: "GitHub設定",
-    github_username: "GitHubユーザー名",
-    username_placeholder: "例: octocat",
-    view_on_github: "GitHubで見る",
-    set_calendar_url: "設定ボタンからカレンダーの埋め込みURLを設定してください",
-    calendar_settings: "Googleカレンダー設定",
-    calendar_settings_desc: "Googleカレンダーの設定から「限定公開の共有可能なリンク」または「埋め込みコード」を取得し、URLを貼り付けてください。",
-    embed_url: "埋め込みURL",
-    display_mode: "表示モード",
-    month: "月 (Month)",
-    week: "週 (Week)",
-    agenda: "スケジュール (Agenda)",
-    reset_widget_sizes_confirm: "全ウィジェットのサイズをリセットしますか？",
-    widget_sizes_reset: "ウィジェットサイズをリセットしました。",
-    fetch_failed: "データの取得に失敗しました",
-    cancel: "キャンセル",
-    no_activity: "No activity",
-    good_job: "Good job!",
-    excellent: "Excellent!",
-    contributions_label: "contributions",
-    day_streak_label: "day streak",
-    longest_streak_label: "longest"
-  },
-  en: {
-    files: "Files",
-    settings: "Settings",
-    add_app: "Add an app",
-    add_widget: "Add a widget",
-    upload_image: "Upload app image",
-    app_name: "App Name",
-    url: "URL",
-    close: "Close",
-    add: "Add",
-    change_position: "Change Position",
-    refresh_dev: "Refresh (Dev)",
-    grid_mode: "Grid Mode",
-    grid_mode_on: "ON",
-    grid_mode_off: "OFF",
-    save: "Save",
-    reset_position: "Reset Position",
-    general: "General",
-    keyboard_shortcuts: "Keyboard Shortcuts",
-    design_style: "Design & Style",
-    security_privacy: "Security & Privacy",
-    account: "Account",
-    debug: "Debug",
-    dev_settings: "Developer Settings",
-    about: "About",
-    user_agent: "User Agent:",
-    language: "Language",
-    settings_button_label: "Settings Button",
-    show_settings_button: "Show",
-    hide_settings_button: "Hide",
-    target_display: "Target Display",
-    select_app_type: "Select app type to add",
-    web_app: "Web App / URL",
-    linux_app: "Linux App",
-    add_linux_app: "Add Linux App",
-    command: "Command",
-    run_in_terminal: "Run in terminal",
-    blur_effect: "Blur Effect",
-    enabled: "Enabled",
-    disabled: "Disabled",
-    dark_mode: "Dark Mode",
-    edit: "Edit",
-    delete: "Delete",
-    hide_widget: "Hide widget",
-    edit_app: "Edit App",
-    edit_linux_app: "Edit Linux App",
-    change_image: "Change Image",
-    add_web_app: "Add App",
-    confirm_delete: "Delete?",
-    confirm_remove_from_folder: "Remove from folder?",
-    confirm_reset_position: "Reset all widget and icon positions?",
-    enter_name_and_url: "Please enter name and URL",
-    enter_name_and_command: "Please enter name and command",
-    launch_failed: "Failed to launch app",
-    enter_folder_name: "Enter folder name",
-    color_scheme: "Color Scheme",
-    delete_all_data: "Delete All Data",
-    confirm_delete_all_data: "Delete all data (apps, folders, settings, etc.)?\nThis action cannot be undone.",
-    data_deleted: "All data has been deleted. Reloading page.",
-    file_shortcut: "File Shortcut",
-    folder_shortcut: "Folder Shortcut",
-    open_failed: "Failed to open file/folder",
-    icon_shape: "Icon Shape",
-    square: "Square",
-    circle: "Circle",
-    window_count: "Window Count (for Virtual Desktops)",
-    apply_restart: "Apply & Restart",
-    confirm_restart: "The app needs to restart to change window count. Restart now?",
-    folder_style: "Folder Style",
-    apply_to_all: "Apply to All",
-    google_login: "Log in to Google",
-    google_login_help: "Please log in if the calendar is not displayed",
-    icon_settings: "Icon Settings",
-    system_settings: "System Settings",
-    data_management: "Data Management",
-    folder_settings: "Folder Settings",
-    folder_name: "Folder Name",
-    background_color: "Background Color",
-    opacity: "Opacity",
-    extract_color_from_image: "Extract color from image",
-    upload_image_generic: "Upload Image",
-    no_music_playing: "No music playing",
-    loading: "Loading...",
-    set_username: "Please set username from settings",
-    github_settings: "GitHub Settings",
-    github_username: "GitHub Username",
-    username_placeholder: "Ex: octocat",
-    view_on_github: "View on GitHub",
-    set_calendar_url: "Please set calendar embed URL from settings",
-    calendar_settings: "Google Calendar Settings",
-    calendar_settings_desc: "Get the 'Secret address in iCal format' or 'Embed code' from Google Calendar settings and paste the URL here.",
-    embed_url: "Embed URL",
-    display_mode: "Display Mode",
-    month: "Month",
-    week: "Week",
-    agenda: "Agenda",
-    reset_widget_sizes_confirm: "Reset all widget sizes?",
-    widget_sizes_reset: "Widget sizes reset.",
-    fetch_failed: "Failed to fetch data",
-    cancel: "Cancel",
-    no_activity: "No activity",
-    good_job: "Good job!",
-    excellent: "Excellent!",
-    contributions_label: "contributions",
-    day_streak_label: "day streak",
-    longest_streak_label: "longest"
-  }
-};
-
-function updateLanguage(lang) {
-  document.querySelectorAll('[data-i18n]').forEach(el => {
-    const key = el.getAttribute('data-i18n');
-    if (translations[lang][key]) {
-      el.textContent = translations[lang][key];
-    }
-  });
-  localStorage.setItem(LS_KEYS.LANGUAGE, lang);
-}
-
-const languageSelector = document.getElementById('language_selector');
-if (languageSelector) {
-  languageSelector.addEventListener('change', (e) => {
-    updateLanguage(e.target.value);
-  });
-  
-  const currentLang = localStorage.getItem(LS_KEYS.LANGUAGE) || 'ja';
-  languageSelector.value = currentLang;
-  updateLanguage(currentLang);
-}
+// Translations and language selector logic have been moved to i18n.js
 
 // ========================================
 // カラースキーム設定
@@ -3216,7 +2954,7 @@ if (applyWindowCountBtn && windowCountSelector) {
     const newCount = parseInt(windowCountSelector.value, 10);
     const lang = getCurrentLanguage();
     
-    if (await showConfirmDialog(translations[lang].confirm_restart)) {
+    if (await showConfirmDialog(i18n.t('confirm_restart'))) {
       if (window.electronAPI && window.electronAPI.setWindowCount) {
         await window.electronAPI.setWindowCount(newCount);
       }
@@ -3286,14 +3024,14 @@ const deleteAllDataBtn = document.getElementById('delete_all_data_btn');
 if (deleteAllDataBtn) {
   deleteAllDataBtn.onclick = async () => {
     const lang = getCurrentLanguage();
-    const confirmMsg = translations[lang].confirm_delete_all_data;
+    const confirmMsg = i18n.t('confirm_delete_all_data');
     
     if (await showConfirmDialog(confirmMsg)) {
       // localStorageの全データを削除
       localStorage.clear();
       
       // 削除完了メッセージを表示してページをリロード
-      await showAlertDialog(translations[lang].data_deleted);
+      await showAlertDialog(i18n.t('data_deleted'));
       location.reload();
     }
   };
@@ -3361,7 +3099,7 @@ if (saveNewAppBtn) {
     
     if (!name || !url) {
       const lang = getCurrentLanguage();
-      await showAlertDialog(translations[lang].enter_name_and_url);
+      await showAlertDialog(i18n.t('enter_name_and_url'));
       return;
     }
     
@@ -4102,496 +3840,7 @@ document.getElementById('close_edit_linuxapp_modal').onclick = () => {
 // ========================================
 // GitHub Contribution ウィジェット
 // ========================================
-
-const githubContributionWidget = document.getElementById('github_contribution_widget');
-const githubGraph = document.getElementById('github_contribution_graph');
-const githubTotalContributions = document.getElementById('github_total_contributions');
-const githubCurrentStreak = document.getElementById('github_current_streak');
-const githubLongestStreak = document.getElementById('github_longest_streak');
-const githubUsernameDisplay = document.getElementById('github_username_display');
-const githubSettingsBtn = document.getElementById('github_settings_btn');
-const githubRefreshBtn = document.getElementById('github_refresh_btn');
-const githubSettingsModal = document.getElementById('github_settings_modal_overlay');
-const githubUsernameInput = document.getElementById('github_username_input');
-const githubYearSelect = document.getElementById('github_year_select');
-const githubDetailsModal = document.getElementById('github_details_modal_overlay');
-const githubDetailsDate = document.getElementById('github_details_date');
-const githubDetailsCount = document.getElementById('github_details_count');
-const githubDetailsLevelText = document.getElementById('github_details_level_text');
-const githubDetailsLevelIcon = document.getElementById('github_details_level_icon');
-const openGitHubActivityBtn = document.getElementById('open_github_activity');
-let currentDetailDate = null;
-
-let githubData = null;
-let currentGitHubYear = 'last';
-
-/**
- * GitHub コントリビューションデータを取得
- * @param {string} username - GitHubユーザー名
- * @returns {Promise<Object>}
- */
-async function fetchGitHubContributions(username) {
-  if (!username) return null;
-  
-  try {
-    // GitHub GraphQL API または スクレイピング用のプロキシサービスを使用
-    // ここでは github-contributions-api を使用
-    const response = await fetch(`https://github-contributions-api.jogruber.de/v4/${username}`);
-    if (!response.ok) {
-      throw new Error('Failed to fetch GitHub data');
-    }
-    const data = await response.json();
-    return data;
-  } catch (error) {
-    console.error('GitHub API error:', error);
-    return null;
-  }
-}
-
-/**
- * コントリビューションレベルを計算（0-4）
- * @param {number} count - コントリビューション数
- * @returns {number}
- */
-function getContributionLevel(count) {
-  if (count === 0) return 0;
-  if (count <= 3) return 1;
-  if (count <= 6) return 2;
-  if (count <= 9) return 3;
-  return 4;
-}
-
-/**
- * ストリーク（連続日数）を計算
- * @param {Array} contributions - コントリビューションデータの配列
- * @returns {{current: number, longest: number}}
- */
-function calculateStreaks(contributions) {
-  if (!contributions || contributions.length === 0) {
-    return { current: 0, longest: 0 };
-  }
-  
-  // 日付順にソート（新しい順）
-  const sorted = [...contributions].sort((a, b) => new Date(b.date) - new Date(a.date));
-  
-  let currentStreak = 0;
-  let longestStreak = 0;
-  let tempStreak = 0;
-  let today = new Date();
-  today.setHours(0, 0, 0, 0);
-  
-  // 現在のストリークを計算
-  for (let i = 0; i < sorted.length; i++) {
-    const date = new Date(sorted[i].date);
-    date.setHours(0, 0, 0, 0);
-    const daysDiff = Math.floor((today - date) / (1000 * 60 * 60 * 24));
-    
-    if (daysDiff === i || (daysDiff === i + 1 && i === 0)) {
-      if (sorted[i].count > 0) {
-        currentStreak++;
-      } else if (daysDiff > 0) {
-        break;
-      }
-    } else {
-      break;
-    }
-  }
-  
-  // 最長ストリークを計算
-  const chronological = [...contributions].sort((a, b) => new Date(a.date) - new Date(b.date));
-  for (const day of chronological) {
-    if (day.count > 0) {
-      tempStreak++;
-      longestStreak = Math.max(longestStreak, tempStreak);
-    } else {
-      tempStreak = 0;
-    }
-  }
-  
-  return { current: currentStreak, longest: longestStreak };
-}
-
-/**
- * GitHubツールチップを表示
- */
-function showGitHubTooltip(targetEl, dateStr, count) {
-  const tooltip = document.getElementById('github_tooltip');
-  if (!tooltip) return;
-  
-  const date = new Date(dateStr);
-  // ユーザーのロケールに合わせて日付をフォーマット
-  const options = { year: 'numeric', month: 'short', day: 'numeric' };
-  const formattedDate = date.toLocaleDateString(undefined, options);
-  
-  const countText = count === 0 ? 'No contributions' : `${count} contribution${count !== 1 ? 's' : ''}`;
-  
-  tooltip.innerHTML = `<div style="font-weight: 600; margin-bottom: 2px;">${countText}</div><div style="color: #ccc;">${formattedDate}</div>`;
-  tooltip.style.display = 'block';
-  
-  const rect = targetEl.getBoundingClientRect();
-  const tooltipRect = tooltip.getBoundingClientRect();
-  
-  // 中央揃え
-  let left = rect.left + (rect.width / 2) - (tooltipRect.width / 2);
-  let top = rect.top - tooltipRect.height - 8;
-  
-  // 画面端の調整
-  if (left < 10) left = 10;
-  if (left + tooltipRect.width > window.innerWidth - 10) {
-    left = window.innerWidth - tooltipRect.width - 10;
-  }
-  
-  tooltip.style.left = `${left}px`;
-  tooltip.style.top = `${top}px`;
-}
-
-/**
- * GitHubツールチップを非表示
- */
-function hideGitHubTooltip() {
-  const tooltip = document.getElementById('github_tooltip');
-  if (tooltip) {
-    tooltip.style.display = 'none';
-  }
-}
-
-/**
- * GitHub詳細モーダルを表示
- */
-function showGitHubDetailsModal(dateStr, count, level) {
-  closeAllModals();
-  if (!githubDetailsModal) return;
-  const lang = getCurrentLanguage();
-  
-  currentDetailDate = dateStr;
-  
-  const date = new Date(dateStr);
-  const options = { year: 'numeric', month: 'long', day: 'numeric', weekday: 'long' };
-  if (githubDetailsDate) {
-    githubDetailsDate.textContent = date.toLocaleDateString(undefined, options);
-  }
-  
-  if (githubDetailsCount) {
-    githubDetailsCount.textContent = `${count} contribution${count !== 1 ? 's' : ''}`;
-  }
-  
-  let levelText = translations[lang].no_activity;
-  let iconName = 'sentiment_neutral';
-  let iconColor = 'var(--on-surface-variant)';
-  
-  if (count > 0) {
-    levelText = translations[lang].good_job;
-    iconName = 'check_circle';
-    iconColor = 'var(--primary-color)';
-  }
-  if (level >= 3) {
-    levelText = translations[lang].excellent;
-    iconName = 'local_fire_department';
-    iconColor = '#ff6d00'; // Orange
-  }
-  
-  if (githubDetailsLevelText) githubDetailsLevelText.textContent = levelText;
-  if (githubDetailsLevelIcon) {
-    githubDetailsLevelIcon.setAttribute('name', iconName);
-    githubDetailsLevelIcon.style.color = iconColor;
-  }
-  
-  githubDetailsModal.style.display = 'flex';
-}
-
-/**
- * GitHubコントリビューショングラフを描画
- * @param {Object} data - GitHubのコントリビューションデータ
- * @param {string|number} year - 表示する年 ('last' または西暦)
- */
-function renderGitHubGraph(data, year = 'last') {
-  if (!githubGraph) return;
-  const lang = getCurrentLanguage();
-  
-  githubGraph.innerHTML = '';
-
-  // グラフエリアでのポインターイベントがウィジェットのドラッグを開始させないようにする
-  githubGraph.onpointerdown = (e) => {
-    e.stopPropagation();
-  };
-  
-  if (!data || !data.contributions) {
-    githubGraph.innerHTML = `
-      <div class="github-error">
-        <m3e-icon name="error"></m3e-icon>
-        <span>${translations[lang].fetch_failed}</span>
-      </div>
-    `;
-    return;
-  }
-  
-  const contributions = data.contributions;
-  let total = 0;
-  let startDate, endDate;
-  
-  if (year === 'last') {
-    // 過去1年分
-    const now = new Date();
-    const todayUTC = new Date(Date.UTC(now.getFullYear(), now.getMonth(), now.getDate()));
-    endDate = todayUTC;
-    startDate = new Date(todayUTC);
-    startDate.setUTCDate(startDate.getUTCDate() - 365);
-
-    // totalを計算
-    if (data.total && typeof data.total.lastYear === 'number') {
-      total = data.total.lastYear;
-    } else {
-      const startStr = startDate.toISOString().split('T')[0];
-      const endStr = endDate.toISOString().split('T')[0];
-      total = contributions.reduce((sum, day) => {
-        if (day.date >= startStr && day.date <= endStr) {
-          return sum + day.count;
-        }
-        return sum;
-      }, 0);
-    }
-  } else {
-    // 特定の年
-    const yearNum = parseInt(year);
-    if (data.total && data.total[year]) {
-      total = data.total[year];
-    } else if (data.years) {
-      const yData = data.years.find(y => y.year === year.toString());
-      if (yData) total = yData.total;
-    }
-    
-    startDate = new Date(Date.UTC(yearNum, 0, 1));
-    endDate = new Date(Date.UTC(yearNum, 11, 31));
-  }
-  
-  // 表示開始日を週の始め（日曜日）に調整
-  const dayOfWeek = startDate.getUTCDay();
-  const displayStartDate = new Date(startDate);
-  displayStartDate.setUTCDate(displayStartDate.getUTCDate() - dayOfWeek);
-
-  // 日付でインデックスを作成
-  const contributionMap = {};
-  contributions.forEach(day => {
-    contributionMap[day.date] = day.count;
-  });
-  
-  // グラフを描画（週ごと列、日ごと行）
-  const fragment = document.createDocumentFragment();
-  let currentDate = new Date(displayStartDate);
-  
-  for (let week = 0; week < 53; week++) {
-    for (let day = 0; day < 7; day++) {
-      const dateStr = currentDate.toISOString().split('T')[0];
-      const count = contributionMap[dateStr] || 0;
-      const level = getContributionLevel(count);
-      
-      const dayEl = document.createElement('div');
-      dayEl.className = 'github-day';
-      dayEl.dataset.level = level;
-      dayEl.dataset.date = dateStr;
-      dayEl.dataset.count = count;
-      
-      dayEl.onmouseenter = () => showGitHubTooltip(dayEl, dateStr, count);
-      dayEl.onmouseleave = hideGitHubTooltip;
-      dayEl.onclick = (e) => {
-        e.stopPropagation();
-        showGitHubDetailsModal(dateStr, count, level);
-      };
-      
-      // 特定の年の場合、その年以外の日は非表示にする
-      if (year !== 'last') {
-        const d = new Date(dateStr);
-        if (d.getUTCFullYear() !== parseInt(year)) {
-          dayEl.style.visibility = 'hidden';
-        }
-      }
-
-      fragment.appendChild(dayEl);
-      currentDate.setUTCDate(currentDate.getUTCDate() + 1);
-    }
-  }
-  
-  githubGraph.appendChild(fragment);
-  
-  // 統計を更新
-  let statsContributions = contributions;
-  if (year !== 'last') {
-    statsContributions = contributions.filter(c => c.date.startsWith(year));
-  }
-  
-  const streaks = calculateStreaks(statsContributions);
-  
-  if (githubTotalContributions) {
-    githubTotalContributions.textContent = total.toLocaleString();
-  }
-  if (githubCurrentStreak) {
-    githubCurrentStreak.textContent = streaks.current;
-  }
-  if (githubLongestStreak) {
-    githubLongestStreak.textContent = streaks.longest;
-  }
-}
-
-/**
- * 年選択肢を生成
- */
-function populateGitHubYearSelect(data) {
-  if (!githubYearSelect) return;
-  
-  githubYearSelect.innerHTML = '';
-  
-  // Last Year
-  const lastOption = document.createElement('option');
-  lastOption.value = 'last';
-  lastOption.textContent = 'Last Year';
-  githubYearSelect.appendChild(lastOption);
-  
-  // Years
-  let years = [];
-  if (data.years) {
-    years = data.years.map(y => y.year);
-  } else if (data.total) {
-    years = Object.keys(data.total).filter(k => k !== 'lastYear');
-  }
-  
-  // 降順ソート
-  years.sort((a, b) => b - a);
-  
-  years.forEach(year => {
-    const option = document.createElement('option');
-    option.value = year;
-    option.textContent = year;
-    githubYearSelect.appendChild(option);
-  });
-  
-  githubYearSelect.value = currentGitHubYear;
-  githubYearSelect.style.display = 'block';
-  
-  githubYearSelect.onchange = (e) => {
-    currentGitHubYear = e.target.value;
-    renderGitHubGraph(githubData, currentGitHubYear);
-  };
-
-  // セレクトボックスでのポインターイベントがウィジェットのドラッグを開始させないようにする
-  githubYearSelect.onpointerdown = (e) => {
-    e.stopPropagation();
-  };
-}
-
-/**
- * GitHubウィジェットを更新
- */
-async function updateGitHubWidget() {
-  const username = localStorage.getItem(LS_KEYS.GITHUB_USERNAME);
-  const lang = getCurrentLanguage();
-  
-  if (!username) {
-    if (githubGraph) {
-      githubGraph.innerHTML = `
-        <div class="github-no-user">
-          <m3e-icon name="person_add"></m3e-icon>
-          <span>${translations[lang].set_username}</span>
-        </div>
-      `;
-    }
-    if (githubUsernameDisplay) {
-      githubUsernameDisplay.textContent = translations[lang].set_username;
-    }
-    return;
-  }
-  
-  if (githubUsernameDisplay) {
-    githubUsernameDisplay.textContent = `@${username}`;
-  }
-  
-  // ローディング表示
-  if (githubGraph) {
-    githubGraph.innerHTML = `
-      <div class="github-loading">
-        <m3e-icon name="hourglass_empty"></m3e-icon>
-        <span>${translations[lang].loading}</span>
-      </div>
-    `;
-  }
-  
-  const data = await fetchGitHubContributions(username);
-  githubData = data;
-  populateGitHubYearSelect(data);
-  renderGitHubGraph(data, currentGitHubYear);
-}
-
-// GitHub更新ボタン
-if (githubRefreshBtn) {
-  githubRefreshBtn.onpointerdown = (e) => e.stopPropagation();
-  githubRefreshBtn.onclick = (e) => {
-    e.stopPropagation();
-    updateGitHubWidget();
-  };
-}
-
-// GitHub設定モーダル
-if (githubSettingsBtn) {
-  // ドラッグイベントとの競合を防ぐ
-  githubSettingsBtn.onpointerdown = (e) => {
-    e.stopPropagation();
-  };
-  
-  githubSettingsBtn.onclick = (e) => {
-    e.stopPropagation();
-    e.preventDefault();
-    closeAllModals();
-    const savedUsername = localStorage.getItem(LS_KEYS.GITHUB_USERNAME) || '';
-    if (githubUsernameInput) {
-      githubUsernameInput.value = savedUsername;
-    }
-    if (githubSettingsModal) {
-      githubSettingsModal.style.display = 'flex';
-    }
-  };
-}
-
-document.getElementById('close_github_settings_modal')?.addEventListener('click', () => {
-  if (githubSettingsModal) {
-    githubSettingsModal.style.display = 'none';
-  }
-});
-
-document.getElementById('save_github_settings')?.addEventListener('click', async () => {
-  const username = githubUsernameInput?.value?.trim();
-  if (username) {
-    localStorage.setItem(LS_KEYS.GITHUB_USERNAME, username);
-  } else {
-    localStorage.removeItem(LS_KEYS.GITHUB_USERNAME);
-  }
-  
-  if (githubSettingsModal) {
-    githubSettingsModal.style.display = 'none';
-  }
-  
-  await updateGitHubWidget();
-});
-
-// GitHub詳細モーダルのイベント
-document.getElementById('close_github_details_modal')?.addEventListener('click', () => {
-  if (githubDetailsModal) githubDetailsModal.style.display = 'none';
-});
-
-openGitHubActivityBtn?.addEventListener('click', () => {
-  if (currentDetailDate) {
-    const username = localStorage.getItem(LS_KEYS.GITHUB_USERNAME);
-    if (username) {
-      const url = `https://github.com/${username}?tab=overview&from=${currentDetailDate}&to=${currentDetailDate}`;
-      window.open(url, '_blank');
-    }
-  }
-  if (githubDetailsModal) githubDetailsModal.style.display = 'none';
-});
-
-// 初期化時にGitHubウィジェットを更新
-setTimeout(updateGitHubWidget, 1000);
-
-// 1時間ごとにGitHubウィジェットを更新
-setInterval(updateGitHubWidget, 60 * 60 * 1000);
+// Moved to widgets/github_widget.js
 
 // ========================================
 // Google Calendar Widget
@@ -4824,7 +4073,7 @@ function createFileShortcutIcon(fileData) {
     const result = await openFileOrFolder(fileData.path);
     if (!result.success) {
       const lang = getCurrentLanguage();
-      await showAlertDialog(translations[lang].open_failed + ': ' + result.error);
+      await showAlertDialog(i18n.t('open_failed') + ': ' + result.error);
     }
   };
   
@@ -4885,7 +4134,7 @@ function createFolderShortcutIcon(folderData) {
     const result = await openFileOrFolder(folderData.path);
     if (!result.success) {
       const lang = getCurrentLanguage();
-      await showAlertDialog(translations[lang].open_failed + ': ' + result.error);
+      await showAlertDialog(i18n.t('open_failed') + ': ' + result.error);
     }
   };
   
@@ -5147,9 +4396,9 @@ document.addEventListener('DOMContentLoaded', () => {
   if (btn) {
     btn.addEventListener('click', async () => {
       const lang = getCurrentLanguage();
-      if (!await showConfirmDialog(translations[lang].reset_widget_sizes_confirm)) return;
+      if (!await showConfirmDialog(i18n.t('reset_widget_sizes_confirm'))) return;
       resetWidgetSizes();
-      await showAlertDialog(translations[lang].widget_sizes_reset);
+      await showAlertDialog(i18n.t('widget_sizes_reset'));
     });
   }
