+++ index.html
@@ -5,8 +5,10 @@
     <link rel="stylesheet" href="./clock.css">
     <link rel="stylesheet" href="./desktop.css">
     <script src="./clock.js" defer></script>
+    <script src="./constants.js" defer></script>
     <script src="./i18n.js" defer></script>
     <script src="./desktop.js" defer></script>
+    <script src="./widgets/github_widget.js" defer></script>
 
     <script type="importmap">
       {
